requirements.txt
# API
fastapi
uvicorn[standard]
httpx

# Modelos y validación
pydantic>=2.0

# Agente (configs YAML)
pyyaml

# UI
streamlit
